Add first try for IconInput
This one went pretty well, although I had to use a React.useState in
order to get the focus ring correct. I suspect that there is a way to do
this in pure CSS, but I couldn't quite figure it out. Maybe a sibling
selector? Actually I'm going to go try that now...